+++ packages/ntask/Sources/NTaskLib/NTaskCommand.swift
@@ -1,7 +1,7 @@
 import ArgumentParser
 
 public enum NTaskVersion {
-    public static let current = "0.3.0"
+    public static let current = "0.3.1"
 }
 
 public struct NTaskCommand: AsyncParsableCommand {
@@ -21,6 +21,8 @@ public struct NTaskCommand: AsyncParsableCommand {
             Get.self,
             Comment.self,
             Review.self,
+            Approve.self,
+            Rework.self,
             Cancel.self,
             Update.self,
             Version.self,